Refactor clean_up_table()

+++ dbt/adapters/athena/impl.py
@@ -2,7 +2,7 @@ from uuid import uuid4
 import agate
 import re
 import boto3
-from pyathena.error import OperationalError
+from botocore.exceptions import ClientError
 
 from dbt.adapters.base import available
 from dbt.adapters.sql import SQLAdapter
@@ -46,13 +46,11 @@ class AthenaAdapter(SQLAdapter):
     def clean_up_partitions(
         self, database_name: str, table_name: str, where_condition: str
     ):
-        logger.info("database_name='{}', table_name='{}', where_condition='{}'", database_name, table_name, where_condition)
-
         # Look up Glue partitions & clean up
         glue_client = boto3.client('glue')
         s3_resource = boto3.resource('s3')
         partitions = glue_client.get_partitions(
-            # CatalogId='awsdatacatalog',  # Using this caused permission error that 'glue:GetPartitions' is required
+            # CatalogId='123456789012', # Need to make this configurable if it is different from default AWS Account ID
             DatabaseName=database_name,
             TableName=table_name,
             Expression=where_condition
@@ -64,6 +62,33 @@ class AthenaAdapter(SQLAdapter):
             if m is not None:
                 bucket_name = m.group(1)
                 prefix = m.group(2)
+                s3_bucket = s3_resource.Bucket(bucket_name)
+                s3_bucket.objects.filter(Prefix=prefix).delete()
+
+    @available
+    def clean_up_table(
+        self, database_name: str, table_name: str
+    ):
+        # Look up Glue partitions & clean up
+        glue_client = boto3.client('glue')
+        try:
+            table = glue_client.get_table(
+                DatabaseName=database_name,
+                Name=table_name
+            )
+        except ClientError as e:
+            if e.response['Error']['Code'] == 'EntityNotFoundException':
+                logger.debug("Table '{}' does not exists - Ignoring", table_name)
+                return
+
+        if table is not None:
+            logger.debug("Deleting table data from'{}'", table["Table"]["StorageDescriptor"]["Location"])
+            p = re.compile('s3://([^/]*)/(.*)')
+            m = p.match(table["Table"]["StorageDescriptor"]["Location"])
+            if m is not None:
+                bucket_name = m.group(1)
+                prefix = m.group(2)
+                s3_resource = boto3.resource('s3')
                 s3_bucket = s3_resource.Bucket(bucket_name)
                 s3_bucket.objects.filter(Prefix=prefix).delete()
